+++ src/software/new_geom/util/circle_tangent_test.cpp
@@ -1,0 +1,3 @@
+#include "software/new_geom/util/circle_tangent.h"
+
+#include <gtest/gtest.h>
